+++ data/format_energy_weights.json
