+++ data/style_conditioning_bank.json
