+++ data/universal_gpp_strength.json
