+++ fightcamp/camp_phases.py
