+++ fightcamp/injury_subs.py
